requirements.txt
praw
pandas
openpyxl
tweepy
pytube
fastapi
Jinja2
yake
python-dotenv
uvicorn
gunicorn
requests
numpy
DateTime
python-multipart
wordcloud
matplotlib
